chore: 메일 디렉토리 구조 변경 및 구글 앱 연동
