+++ src/main/java/com/intellimarket/common/service/MailService.java
@@ -1,6 +1,7 @@
-package com.intellimarket.common.mail.service;
+package com.intellimarket.common.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.mail.MailException;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
@@ -12,11 +13,12 @@ import com.intellimarket.common.exception.CommonException;
 public class MailService {
 	@Autowired
 	private JavaMailSender mailSender;
+	@Value("${mail.username}") String username;
 	
 	public void sendMail(String toEmail, String subject, String text) {
 		SimpleMailMessage message = new SimpleMailMessage();
 		
-		message.setFrom(""); 								// 발신자
+		message.setFrom(username); 							// 발신자
 		message.setTo(toEmail); 							// 수신자
 		message.setSubject(subject); 						// 메일 제목
 		message.setText(text); 								// 메일 내용
@@ -25,7 +27,7 @@ public class MailService {
 			mailSender.send(message);
 		} catch (MailException e) {
 			e.printStackTrace();
-			throw new CommonException("인증 메일 전송에 실패했습니다.");
+			throw new CommonException("메일 전송에 실패했습니다.");
 		}
 	}
 }
